SONARJAVA-485 Prevent creation of findbugs directory
Fix bad null pointer exception.

+++ sonar-findbugs-plugin/src/main/java/org/sonar/plugins/findbugs/FindbugsConfiguration.java
@@ -150,8 +150,12 @@ public class FindbugsConfiguration implements BatchExtension {
    * Invoked by PicoContainer to remove temporary files.
    */
   public void stop() {
-    jsr305Lib.delete();
-    annotationsLib.delete();
+    if (jsr305Lib != null) {
+      jsr305Lib.delete();
+    }
+    if (annotationsLib != null) {
+      annotationsLib.delete();
+    }
   }
 
   private File copyLib(String name) {
